Move `AddTag` to separate file
